*: move "etcdserver/api/v3compactor"
