+++ etcdserver/api/v3compactor/periodic.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compactor
+package v3compactor
 
 import (
 	"context"
